Added APIs to specify the ending position
This change also changed the way start position works. Instead of informing the ICU library about the starting index, this change will move the pointer position of the string buffer to the starting index. This will allow anchor (i.e. \A) to recognize the first character in the specified range as the first character that appeared in the string. This is the same as how NSRegularExpresion behaves.